refactor: moved the initialization specs onto the machine learning specs from the parameter server
